feat!: move assertion declarations to expect package
