+++ .eslintrc
@@ -4,6 +4,7 @@
     "no-only-tests/no-only-tests": "off",
     // prefer global Buffer to not initialize the whole module
     "n/prefer-global/buffer": "off",
+    "@typescript-eslint/no-invalid-this": "off",
     "no-restricted-imports": [
       "error",
       {
