+++ packages/expect/src/state.ts
@@ -1,8 +1,8 @@
-import type { MatcherState } from './types'
+import type { ExpectStatic, MatcherState } from './types'
 import { GLOBAL_EXPECT, JEST_MATCHERS_OBJECT, MATCHERS_OBJECT } from './constants'
 
 if (!Object.prototype.hasOwnProperty.call(globalThis, MATCHERS_OBJECT)) {
-  const globalState = new WeakMap<Vi.ExpectStatic, MatcherState>()
+  const globalState = new WeakMap<ExpectStatic, MatcherState>()
   const matchers = Object.create(null)
   Object.defineProperty(globalThis, MATCHERS_OBJECT, {
     get: () => globalState,
@@ -16,13 +16,13 @@ if (!Object.prototype.hasOwnProperty.call(globalThis, MATCHERS_OBJECT)) {
   })
 }
 
-export function getState<State extends MatcherState = MatcherState>(expect: Vi.ExpectStatic): State {
+export function getState<State extends MatcherState = MatcherState>(expect: ExpectStatic): State {
   return (globalThis as any)[MATCHERS_OBJECT].get(expect)
 }
 
 export function setState<State extends MatcherState = MatcherState>(
   state: Partial<State>,
-  expect: Vi.ExpectStatic,
+  expect: ExpectStatic,
 ): void {
   const map = (globalThis as any)[MATCHERS_OBJECT]
   const current = map.get(expect) || {}
