+++ packages/expect/src/types.ts
@@ -9,6 +9,7 @@ import type { use as chaiUse } from 'chai'
  */
 
 import type { Formatter } from 'picocolors/types'
+import type { Constructable } from '@vitest/utils'
 import type { diff, getMatcherUtils, stringify } from './jest-matcher-utils'
 
 export type FirstFunctionArgument<T> = T extends (arg: infer A) => unknown ? A : never
@@ -96,3 +97,97 @@ export interface RawMatcherFn<T extends MatcherState = MatcherState> {
 }
 
 export type MatchersObject<T extends MatcherState = MatcherState> = Record<string, RawMatcherFn<T>>
+
+export interface ExpectStatic extends Chai.ExpectStatic, AsymmetricMatchersContaining {
+  <T>(actual: T, message?: string): Assertion<T>
+
+  extend(expects: MatchersObject): void
+  assertions(expected: number): void
+  hasAssertions(): void
+  anything(): any
+  any(constructor: unknown): any
+  getState(): MatcherState
+  setState(state: Partial<MatcherState>): void
+  not: AsymmetricMatchersContaining
+}
+
+export interface AsymmetricMatchersContaining {
+  stringContaining(expected: string): any
+  objectContaining<T = any>(expected: T): any
+  arrayContaining<T = unknown>(expected: Array<T>): any
+  stringMatching(expected: string | RegExp): any
+}
+
+export interface JestAssertion<T = any> extends jest.Matchers<void, T> {
+  // Jest compact
+  toEqual<E>(expected: E): void
+  toStrictEqual<E>(expected: E): void
+  toBe<E>(expected: E): void
+  toMatch(expected: string | RegExp): void
+  toMatchObject<E extends {} | any[]>(expected: E): void
+  toContain<E>(item: E): void
+  toContainEqual<E>(item: E): void
+  toBeTruthy(): void
+  toBeFalsy(): void
+  toBeGreaterThan(num: number | bigint): void
+  toBeGreaterThanOrEqual(num: number | bigint): void
+  toBeLessThan(num: number | bigint): void
+  toBeLessThanOrEqual(num: number | bigint): void
+  toBeNaN(): void
+  toBeUndefined(): void
+  toBeNull(): void
+  toBeDefined(): void
+  toBeInstanceOf<E>(expected: E): void
+  toBeCalledTimes(times: number): void
+  toHaveLength(length: number): void
+  toHaveProperty<E>(property: string | (string | number)[], value?: E): void
+  toBeCloseTo(number: number, numDigits?: number): void
+  toHaveBeenCalledTimes(times: number): void
+  toHaveBeenCalled(): void
+  toBeCalled(): void
+  toHaveBeenCalledWith<E extends any[]>(...args: E): void
+  toBeCalledWith<E extends any[]>(...args: E): void
+  toHaveBeenNthCalledWith<E extends any[]>(n: number, ...args: E): void
+  nthCalledWith<E extends any[]>(nthCall: number, ...args: E): void
+  toHaveBeenLastCalledWith<E extends any[]>(...args: E): void
+  lastCalledWith<E extends any[]>(...args: E): void
+  toThrow(expected?: string | Constructable | RegExp | Error): void
+  toThrowError(expected?: string | Constructable | RegExp | Error): void
+  toReturn(): void
+  toHaveReturned(): void
+  toReturnTimes(times: number): void
+  toHaveReturnedTimes(times: number): void
+  toReturnWith<E>(value: E): void
+  toHaveReturnedWith<E>(value: E): void
+  toHaveLastReturnedWith<E>(value: E): void
+  lastReturnedWith<E>(value: E): void
+  toHaveNthReturnedWith<E>(nthCall: number, value: E): void
+  nthReturnedWith<E>(nthCall: number, value: E): void
+}
+
+// eslint-disable-next-line @typescript-eslint/prefer-ts-expect-error
+// @ts-ignore build namespace conflict
+type VitestAssertion<A, T> = {
+  [K in keyof A]: A[K] extends Chai.Assertion
+    ? Assertion<T>
+    : A[K] extends (...args: any[]) => any
+      ? A[K] // not converting function since they may contain overload
+      : VitestAssertion<A[K], T>
+} & ((type: string, message?: string) => Assertion)
+
+type Promisify<O> = {
+  [K in keyof O]: O[K] extends (...args: infer A) => infer R
+    ? O extends R
+      ? Promisify<O[K]>
+      : (...args: A) => Promise<R>
+    : O[K]
+}
+
+export interface Assertion<T = any> extends VitestAssertion<Chai.Assertion, T>, JestAssertion<T> {
+  toBeTypeOf(expected: 'bigint' | 'boolean' | 'function' | 'number' | 'object' | 'string' | 'symbol' | 'undefined'): void
+  toHaveBeenCalledOnce(): void
+  toSatisfy<E>(matcher: (value: E) => boolean, message?: string): void
+
+  resolves: Promisify<Assertion<T>>
+  rejects: Promisify<Assertion<T>>
+}
