+++ packages/vitest/src/integrations/chai/index.ts
@@ -5,21 +5,22 @@ import './setup'
 import type { Test } from '@vitest/runner'
 import { getCurrentTest } from '@vitest/runner'
 import { GLOBAL_EXPECT, getState, setState } from '@vitest/expect'
+import type { Assertion, ExpectStatic } from '@vitest/expect'
 import type { MatcherState } from '../../types/chai'
 import { getCurrentEnvironment, getFullName } from '../../utils'
 
 export function createExpect(test?: Test) {
-  const expect = ((value: any, message?: string): Vi.Assertion => {
+  const expect = ((value: any, message?: string): Assertion => {
     const { assertionCalls } = getState(expect)
     setState({ assertionCalls: assertionCalls + 1 }, expect)
-    const assert = chai.expect(value, message) as unknown as Vi.Assertion
+    const assert = chai.expect(value, message) as unknown as Assertion
     const _test = test || getCurrentTest()
     if (_test)
       // @ts-expect-error internal
-      return assert.withTest(_test) as Vi.Assertion
+      return assert.withTest(_test) as Assertion
     else
       return assert
-  }) as Vi.ExpectStatic
+  }) as ExpectStatic
   Object.assign(expect, chai.expect)
 
   expect.getState = () => getState<MatcherState>(expect)
